pracker_site/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-js-asset==2.0.0
django-mptt==0.14.0
django-truncate==0.1
psycopg2-binary==2.9.5
python-slugify==7.0.0
sqlparse==0.4.3
text-unidecode==1.3
tornado==6.2
tzdata==2022.7
watchdog==2.1.9
websocket-client==1.4.2
